All files: group functions and classes into more classes
Some classes are very long and can be divided into multiple
classes.

Creating more classes to further group similar functions allows
higher encapsulation and code quality.

Let's
* Add classes UiManager, Parser and Storage
* Move functions that handled display from Bob.java to UiManager
* Distribute functions in TaskManager to Parser and Storage

Aside from Parser and Storage mentioned in the assignment,
UiManager works similarly to Ui and TaskManager works similarly
to TaskList.
